fix(folder): add custom fields as its own folder
