+++ standalone/src/models/custom-fields/custom-fields/custom-fields-draft/presets/with-string-field.ts
