+++ standalone/src/models/custom-fields/custom-fields/index.ts
@@ -1,0 +1,12 @@
+import { RestModelBuilder, GraphqlModelBuilder } from './builders';
+import * as modelPresets from './presets';
+
+export const CustomFieldsRest = {
+  random: RestModelBuilder,
+  presets: modelPresets.restPresets,
+};
+
+export const CustomFieldsGraphql = {
+  random: GraphqlModelBuilder,
+  presets: modelPresets.graphqlPresets,
+};
