+++ standalone/src/models/custom-fields/custom-fields/presets/with-string-field.spec.ts
